MSM/Serialization: test completeness
